+++ src/components/shared/contracts/ContractSummary.js
@@ -3,9 +3,16 @@ import { Link } from "react-router-dom";
 import { Chip, IconButton } from "@material-ui/core";
 import InfoIcon from "@material-ui/icons/Info";
 import { useTranslation } from "react-i18next";
+import PluginRegistry from "../../core/PluginRegistry";
 
 const ContractSummary = ({ orgUnit, contract }) => {
   const { t } = useTranslation();
+  const program = PluginRegistry.extension("contracts.program")
+
+  const options = program.programStages.flatMap(ps => ps.programStageDataElements).map(psde => psde.dataElement).flatMap(de => de.optionSet?.options).filter(o => o)
+
+  const optionsByCode = _.keyBy(options, o => o.code)
+
   return (
     <div>
       {t("dataEntry.contractFrom")} <code>{contract.startPeriod}</code> {t("dataEntry.contractTo")}{" "}
@@ -21,7 +28,7 @@ const ContractSummary = ({ orgUnit, contract }) => {
         </IconButton>
       </Link>{" "}
       {contract.codes.map((c, index) => (
-        <Chip key={c + "_" + index} label={c} style={{ margin: "5px" }} />
+        <Chip key={c + "_" + index} title={c} label={optionsByCode[c] ? optionsByCode[c].name : c} style={{ margin: "5px" }} />
       ))}
     </div>
   );